Move over dateAndTimeInputsAreValidDates
