+++ server/form-pages/apply/type-of-ap/pipeReferral.ts
@@ -1,7 +1,8 @@
 import type { YesOrNo, ObjectWithDateParts } from 'approved-premises'
 
 import TasklistPage from '../../tasklistPage'
-import { dateAndTimeInputsAreValidDates, dateIsBlank } from '../../../utils/utils'
+import { dateIsBlank } from '../../../utils/utils'
+import { dateAndTimeInputsAreValidDates } from '../../../utils/dateUtils'
 
 export default class PipeReferral implements TasklistPage {
   name = 'pipe-referral'
