+++ server/utils/utils.test.ts
@@ -1,12 +1,6 @@
 import type { ObjectWithDateParts } from 'approved-premises'
 import { SessionDataError } from './errors'
-import {
-  convertToTitleCase,
-  initialiseName,
-  dateAndTimeInputsAreValidDates,
-  retrieveQuestionResponseFromSession,
-  dateIsBlank,
-} from './utils'
+import { convertToTitleCase, initialiseName, retrieveQuestionResponseFromSession, dateIsBlank } from './utils'
 
 describe('convert to title case', () => {
   it.each([
@@ -34,32 +28,6 @@ describe('initialise name', () => {
     ['Double barrelled', 'Robert-John Smith-Jones-Wilson', 'R. Smith-Jones-Wilson'],
   ])('%s initialiseName(%s, %s)', (_: string, a: string, expected: string) => {
     expect(initialiseName(a)).toEqual(expected)
-  })
-})
-
-describe('dateAndTimeInputsAreValidDates', () => {
-  it('returns true when the date is valid', () => {
-    const obj: ObjectWithDateParts<'date'> = {
-      'date-year': '2022',
-      'date-month': '12',
-      'date-day': '11',
-    }
-
-    const result = dateAndTimeInputsAreValidDates(obj, 'date')
-
-    expect(result).toEqual(true)
-  })
-
-  it('returns false when the date is invalid', () => {
-    const obj: ObjectWithDateParts<'date'> = {
-      'date-year': '99',
-      'date-month': '99',
-      'date-day': '99',
-    }
-
-    const result = dateAndTimeInputsAreValidDates(obj, 'date')
-
-    expect(result).toEqual(false)
   })
 })
